Add pkcs12_keystore type & provider
This provider provides an alternative to the export::pkcs12 define.
The main difference being that this manages the actual contents of the
keystore and doesn't only trigger whan an input changes.

I initially wanted to implement this purely in Ruby, but because AIO
packages ship their own openssl this could cause weird issues when it
differs too much from system openssl.