Add pkcs12_keystore type & provider
This is a completely ruby native implementation, so this does not use
the `openssl` command like the other providers. It's also more powerful
than the export::pkcs12 function in the sense that it can actually
properly manage (and track changes to) a keystore instead of purely
creating it with execs.